+++ src/py/flwr/common/record/__init__.py
@@ -19,7 +19,7 @@ from .configsrecord import ConfigsRecord
 from .conversion_utils import array_from_numpy
 from .metricsrecord import MetricsRecord
 from .parametersrecord import Array, ParametersRecord
-from .recordset import RecordSet
+from .recorddict import RecordSet
 
 __all__ = [
     "Array",
